Added grunt-config Added configuration file for server
