+++ Gruntfile.js
@@ -1,98 +1,24 @@
 module.exports = function(grunt) {
     'use strict';
-    grunt.initConfig({
-        pkg: grunt.file.readJSON('package.json'),
-
-        docco: {
-            debug: {
-                src: ['test/**/*.js'],
-                options: {
-                    output: 'docs/'
-                }
-            }
-        },
-
-        express: {
-            dev: {
-                options: {
-                    script: 'gui-resources/scripts/js/server.js',
-                    port: 3000
-                }
-            }
-        },
-
-        jshint: {
-            options: {
-                ignores: ['gui-resources/scripts/js/bower_components/**']
-            },
-            own: {
-                src: ['Gruntfile.js', 'gui-resources/scripts/js/**/*.js', '!gui-resources/scripts/js/core/**/*.js']
-            },
-            libs: {
-                src: ['gui-resources/scripts/js/core/**/*.js']
-            },
-            all: {
-                src: ['Gruntfile.js', 'gui-resources/scripts/js/**/*.js']
-            }
-        },
-
-        open: {
-            dev: {
-                path: 'http://localhost:<%= express.dev.options.port %>'
-            }
-        },
-
-        requirejs: {
-            compile: {
-                options: {
-                    mainConfigFile: 'gui-resources/scripts/js/main.js',
-                    baseUrl: 'gui-resources/scripts/js/',
-                    name: 'main',
-                    out: 'build/app.min.js',
-                    findNestedDependencies: true
-                }
-            }
-        },
-
-        watch: {
-            express: {
-                files: ['gui-resources/scripts/js/**/*.js'],
-                tasks: ['express:dev'],
-                options: {
-                    // According to express docu, 'spawn: false' is needed for the
-                    // server to reload, but if we use it the browser is reloaded
-                    // before the server is ready
-                    //spawn: false,
-                    livereload: true
-                }
-            },
-            'hint-own': {
-                files: ['<%= jshint.own.src %>'],
-                tasks: ['jshint:own']
-            },
-            'hint-libs': {
-                files: ['<%= jshint.libs.src %>'],
-                tasks: ['jshint:libs']
-            },
-            'hint-all': {
-                files: ['<%= jshint.all.src %>'],
-                tasks: ['jshint:all']
-            }
-        }
-    });
 
     // Load all grunt task declared in package.json
     require('load-grunt-tasks')(grunt);
+    
+    // var server =  grunt.option('server') || 'http://localhost:9000',
+    //     port = server.
 
-    grunt.registerTask('install-bower', 'install the frontend dependencies', function() {
-            var exec = require('child_process').exec;
-            var cb = this.async();
-            exec('bower install', {cwd: '.'}, function(err, stdout, stderr) {
-                console.log(stdout);
-                cb();
-            });
-        });
+    // Define basic configuration
+    var config = grunt.file.readJSON('./config.json');
+    config.pkg = grunt.file.readJSON('./package.json');
 
+    // Load the basic configuration and additional from tasks/options
+    require('load-grunt-config')(grunt, {
+        config: config,
+        configPath: require('path').join(process.cwd(), 'tasks', 'options'),
+        init: true
+    });
+
+    // Install the jshint pre-commit hook
     grunt.registerTask('install-hook', function () {
             var fs = require('fs');
