+++ tasks/options/express.js
@@ -1,0 +1,9 @@
+module.exports = {
+    dev: {
+        options: {
+            script: '<%= dir.script  %>/server.js',
+            port: '<%= server.port %>',
+            livereload: true
+        }
+    }
+};
