+++ tasks/options/open.js
@@ -1,0 +1,5 @@
+module.exports = {
+    dev: {
+        path: '<%= server.protocol %><%= server.hostname %>:<%= server.port %>'
+    }
+};
