+++ tasks/options/requirejs.js
@@ -1,0 +1,11 @@
+module.exports = {
+    compile: {
+        options: {
+            mainConfigFile: '<%= dir.script  %>/main.js',
+            baseUrl: '<%= dir.script  %>/',
+            name: 'main',
+            out: '<%= dir.build  %>/app.min.js',
+            findNestedDependencies: true
+        }
+    }
+}
